modularizacion con errores
