+++ src/Controlers/GetAllPlatforms.js
@@ -1,0 +1,8 @@
+const { platform } = require("../db");
+
+const GetALLPlatforms = async () => {
+    const plataformas = await platform.findAll();
+    return plataformas.map(p => p.name);
+}
+
+module.exports = GetALLPlatforms;
