+++ src/routes/index.js
@@ -1,7 +1,7 @@
 const { Router } = require('express') 
-const VGrouter = require('./VGrouter.js') 
-const UserRouter = require('./UserRouter.js')
-const Favorites = require('./Favorites.js')
+const VGrouter = require('./Grouter Videogames/VGrouter') 
+const UserRouter = require('./Routers users/UserRouter')
+const Favorites = require('./Favorite games/Favorites.js')
 const getplatformshandler = require('../handlers/getplatformshandler.js')
 const getGeresHandler = require('../handlers/getGeresHandler.js')
